Update to work on OSX
The sed scripts were written in a manner that looks to assume GNU sed.
OSX, however, has BSD sed, which has a few differences. In order to use
the named character classes, the -E flag is needed. The \n in the s///
expression does not turn into a newline, instead using shell expansion
is needed ($'...\\\n....', which produces a \ followed by a newline).
The + shouldn't be escaped. I don't know if these updated scripts will
continue to work on GNU sed :(